+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 **Weather sources**
 - [ATMO France] fix for no pollen in Paris, Marseille and Lyon
 - [CWA] Fix failure to refresh on Android < 14
+- [Open Meteo] Update weather model list (adds KMA, migrate deprecated models)
 
 **Translations**
 - Translations updated